should set original for scenario format
