+++ app/views/editor/campaign/CampaignEditorView.js
@@ -32,6 +32,7 @@ const PatchesView = require('views/editor/PatchesView')
 const RevertModal = require('views/modal/RevertModal')
 const modelDeltas = require('lib/modelDeltas')
 const globalVar = require('core/globalVar')
+const ScenarioNode = require('views/editor/ai-scenario/AIScenarioNode')
 require('vendor/scripts/jquery-ui-1.11.1.custom')
 require('vendor/styles/jquery-ui-1.11.1.custom.css')
 
@@ -378,7 +379,8 @@ module.exports = (CampaignEditorView = (function () {
           campaigns: CampaignsNode,
           campaign: CampaignNode,
           achievement: AchievementNode,
-          rewards: RewardsNode
+          rewards: RewardsNode,
+          scenario: ScenarioNode,
         },
         supermodel: this.supermodel
       }
